+++ demisto_sdk/commands/pre_commit/pre_commit_command.py
@@ -137,15 +137,15 @@ class PreCommitRunner:
 
     def _get_hooks(self, pre_commit_config: dict) -> dict:
         hooks = {}
-        for repo in pre_commit_config["repos"]:
-            for hook in repo["hooks"]:
+        for i, repo in enumerate(pre_commit_config["repos"]):
+            for j, hook in enumerate(repo["hooks"]):
+                if Hook.get_property(hook, self.mode, "skip"):
+                    del pre_commit_config["repos"][i]["hooks"][j]
+                    continue
                 hooks[hook["id"]] = {
                     "repo": repo,
                     "hook": hook,
                 }
-                # if the hook has a skip key, we add it to the SKIPPED_HOOKS set
-                if hook.pop("skip", None):
-                    SKIPPED_HOOKS.add(hook["id"])
         return hooks
 
     def exclude_python2_of_non_supported_hooks(self) -> None:
@@ -170,7 +170,8 @@ class PreCommitRunner:
             else:
                 hook["hook"]["exclude"] = join_files_string
 
-    def prepare_hooks(self, hooks: dict, run_docker_hooks: bool, dry_run: bool) -> None:
+    def prepare_hooks(self, run_docker_hooks: bool, dry_run: bool) -> None:
+        hooks = self.hooks
         kwargs = {
             "mode": self.mode,
             "all_files": self.all_files,
@@ -261,9 +262,11 @@ class PreCommitRunner:
                 f"Running pre-commit with Python {python_version} on {changed_files_string}"
             )
 
-        self.prepare_hooks(self.hooks, run_docker_hooks, dry_run)
+        self.prepare_hooks(run_docker_hooks, dry_run)
         if self.all_files:
-            self.precommit_template["exclude"] += f"|{join_files(exclude_files or {})}"
+            self.precommit_template[
+                "exclude"
+            ] += f"|{join_files(exclude_files or set())}"
         else:
             self.precommit_template["files"] = join_files(self.files_to_run)