+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed an issue where **run-playbook** command didn't work if the *url* argument was not provided.
 * Fixed an issue where **validate** command failed on valid complex layout rules and triggers.
 * Fixed an issue where *validate* command failed with release notes files for assets modeling rules folder.
+* Added support for `skip` property in **pre-commit** hooks.
 
 ## 1.24.0
 * Fixed an issue where the error was not clear when trying to retrieve the server version.
@@ -14,7 +15,6 @@
 * Fixed an issue where **run-playbook** command did not work.
 * Fixed an issue in **setup-env** command where the virtual environment failed to set up.
 * Fixed an issue in **pre-commit** command where `False` properties were deleted.
-* Fixed an issue in **coverage-analyze** command where the `report_dir` does not exist.
 * Added support for json5 file, allowing to write files and read files from specific git branches, local file system, or from any remote api .
 * Fixed an issue in **upload** command where the `marketplace` field was not taken into consideration when uploading single content-items.
 * Added support for *Assets Modeling Rule* new content item in all `demisto-sdk` commands.